feat: helidon implementation

+++ helidon/src/main/java/io/helidon/labs/dto/StoreFruitPriceDto.java
@@ -1,7 +1,6 @@
 package io.helidon.labs.dto;
 
 import io.helidon.json.binding.Json;
-import java.math.BigDecimal;
 
 @Json.Entity
-public record StoreFruitPriceDto(StoreDto store, BigDecimal price) {}
+public record StoreFruitPriceDto(StoreDto store, float price) {}
